refactor: 🚚 Rename utilities to diffsync to clarify use
As these functions are for both Infoblox and Nautobot the name diffsync makes more sense to align with use. As we expand these moving to individual files for each would be warranted. This makes the naming with the util folder more sensible.

Rename utilities to diffsync so paths are changed and location of functions needs to be updated.
